backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.11.0
celery==5.5.3
redis==6.4.0
python-dotenv==1.1.1
pydantic==2.12.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2
email-validator==2.1.0
